+++ src/main/java/dinkplugin/DinkPluginConfig.java
@@ -415,6 +415,20 @@ public interface DinkPluginConfig extends Config {
     }
 
     @ConfigItem(
+        keyName = SettingsManager.DYNAMIC_IMPORT_CONFIG_KEY,
+        name = "Dynamic Config URL",
+        description = "Synchronizes your Dink configuration with the specified URL.<br/>" +
+            "Whenever Dink starts, it imports the config offered by the URL.<br/>" +
+            "The config applies to all webhooks, so ensure you trust this URL.<br/>" +
+            "Only one URL is supported",
+        position = 1017,
+        section = advancedSection
+    )
+    default String dynamicConfigUrl() {
+        return "";
+    }
+
+    @ConfigItem(
         keyName = "discordWebhook", // do not rename; would break old configs
         name = "Primary Webhook URLs",
         description = "The default webhook URL to send notifications to, if no override is specified.<br/>" +
@@ -952,13 +966,26 @@ public interface DinkPluginConfig extends Config {
     }
 
     @ConfigItem(
+        keyName = "lootRarityValueIntersection",
+        name = "Require both Rarity and Value",
+        description = "Whether items must exceed <i>both</i> the Min Value AND Rarity thresholds to be notified.<br/>" +
+            "Does not apply to drops where Dink lacks rarity data.<br/>" +
+            "Currently only impacts NPC drops",
+        position = 39,
+        section = lootSection
+    )
+    default boolean lootRarityValueIntersection() {
+        return false;
+    }
+
+    @ConfigItem(
         keyName = "lootNotifMessage",
         name = "Notification Message",
         description = "The message to be sent through the webhook.<br/>" +
             "Use %USERNAME% to insert your username<br/>" +
             "Use %LOOT% to insert the loot<br/>" +
             "Use %SOURCE% to show the source of the loot",
-        position = 39,
+        position = 40,
         section = lootSection
     )
     default String lootNotifyMessage() {